+++ src/AnnotationReader.php
@@ -302,6 +302,7 @@ class AnnotationReader
      * @return T|null
      *
      * @throws AnnotationException
+     * @throws ClassNotFoundException
      *
      * @template T of object
      */